refactor(core): reorganize tools/ into domain subfolders + __tests__/ convention
tools/ was a flat directory with 21 files. Reorganized into four
domain-aligned subfolders:

  tools/auth/       — detection, exploration, validation, providers, gap builders
  tools/explorers/  — browser launch, Playwright/Cypress crawlers, factory, interface
  tools/scoring/    — gap engine, automation maturity, public surface
  tools/repo/       — repo scanner, framework detector

All test files across the entire src/ tree are moved into colocated
__tests__/ directories — a consistent convention that keeps source
listings clean while keeping tests discoverable next to their subject.

No behavior change, no public API change, no schema change. Every
import path updated. package.json test script paths updated. Build and
all 72 tests pass.
