+++ packages/core/src/phases/observe.ts
@@ -1,8 +1,8 @@
 import type { HarnessConfig } from '../schemas/config.schema.js';
 import { RouteInventorySchema, type RouteInventory } from '../schemas/route-inventory.schema.js';
 import { RepoAnalysisSchema, type RepoAnalysis } from '../schemas/repo-analysis.schema.js';
-import { createExplorer } from '../tools/explorer-factory.js';
-import { scanRepo } from '../tools/repo-scanner.js';
+import { createExplorer } from '../tools/explorers/factory.js';
+import { scanRepo } from '../tools/repo/scanner.js';
 import { StateManager } from '../harness/state-manager.js';
 import { logDecision } from '../harness/decision-logger.js';
 import type { RunArtifactsOptions } from '../harness/run-options.js';
