+++ packages/core/src/tools/auth/oauth-providers.ts
