+++ packages/core/src/tools/auth/user-providers.ts
