+++ packages/core/src/tools/explorers/browser.ts
